+++ requirements.txt
@@ -1,1 +1,2 @@
 gradio
+mdtex2html
